Fixed arm and arm64 images

+++ pwn-chal/docker-entrypoint-arm.sh
@@ -1,17 +1,70 @@
 #!/bin/bash
 # Author: @TheFlash2k
 
+# self-yeet
+rm -- "$0"
+
 DEFAULT_PORT=8000
 DEFAULT_CHAL_NAME="chal"
 DEFAULT_BASE="ynetd"
 DEFAULT_LOG_FILE="/var/log/chal.log"
 DEFAULT_START_DIR="/app"
 DEFAULT_FLAG_FILE="/app/flag.txt"
+DEFAULT_REDIRECT_STDERR="y"
+DEFAULT_CONN_TIME="30"
+
+function debug() { [ ! -z "$DEBUG" ] && echo -e "\e[32m[*]\e[0m $1"; }
+function info() { echo -e "\e[36m[i]\e[0m $1"; }
+function error() { echo -e "\e[31m[x]\e[0m $1"; exit 1; }
+function warn() { echo -e "\e[33m[!]\e[0m $1"; }
+function set_default() {
+    # Takes a variable name as a parameter
+    # Returns the already set value if value exists otherwise sets to DEFAULT_variableName
+    local var="$1"
+    local __="DEFAULT_$var"
+    local default="${!__}"
+
+    if [ -z "${!var}" ]; then
+        eval "$var=\$default"
+    else
+        eval "$var=\${!var}"
+    fi
+    echo -n "${!var}"
+}
+
+function invalid() {
+    local opt="$1"
+    error "$opt is set to ${!opt}. Can only be ($2)";
+}
+
+[ ! -z "$DEBUG" ] && debug "Debugging is enabled!"
+
+# Check the variables, if not exists, set to default
+PORT=$(set_default "PORT")
+CHAL_NAME=$(set_default "CHAL_NAME")
+BASE=$(set_default "BASE")
+LOG_FILE=$(set_default "LOG_FILE")
+START_DIR=$(set_default "START_DIR")
+FLAG_FILE=$(set_default "FLAG_FILE")
+REDIRECT_STDERR=$(set_default "REDIRECT_STDERR")
+CONN_TIME=$(set_default "CONN_TIME")
+
+debug "PORT=$PORT"
+debug "CHAL_NAME=$CHAL_NAME"
+debug "BASE=$BASE"
+debug "LOG_FILE=$LOG_FILE"
+debug "START_DIR=$START_DIR"
+debug "FLAG_FILE=$FLAG_FILE"
+debug "REDIRECT_STDERR=$REDIRECT_STDERR"
+debug "CONN_TIME=$CONN_TIME"
+
+# Check if REDIRECT_STDERR is y/n
+shopt -s nocasematch
+[[ "$REDIRECT_STDERR"  != "y" && "$REDIRECT_STDERR" != "n" ]] && invalid "REDIRECT_STDERR" "y/n"
+shopt -u nocasematch
 
 # Check if root:
-if [ "$EUID" -eq 0 ]; then
-    chown -R root:ctf-player /app/
-fi
+[ "$EUID" -eq 0 ] &&  chown -R root:ctf-player /app/
 
 if [ -z "$OVERRIDE_USER" ]; then
     RUN_AS="ctf-player"
@@ -20,43 +73,36 @@ else
     if id "$OVERRIDE_USER" >/dev/null 2>&1; then
         RUN_AS="$OVERRIDE_USER"
     else
+        warn "User $OVERRIDE_USER user doesn't exist. Defaulting to root."
         RUN_AS="root"
     fi
 fi
 
-[ -z "$PORT" ] && PORT="$DEFAULT_PORT"
-[ -z "$CHAL_NAME" ] && CHAL_NAME="$DEFAULT_CHAL_NAME"
-[ -z "$BASE" ] && BASE="$DEFAULT_BASE"
-[ -z "$LOG_FILE" ] && LOG_FILE="$DEFAULT_LOG_FILE"
-[ -z "$START_DIR" ] && START_DIR="$DEFAULT_START_DIR"
-[ -z "$FLAG_FILE" ] && FLAG_FILE="$DEFAULT_FLAG_FILE"
-
-if [ "$BASE" != "ynetd" ] && [ "$BASE" != "socat" ]; then
-    echo "Invalid utility: $BASE. Can only use ynetd or socat."
-    exit 1
-fi
-
-if [ ! -f "/app/$CHAL_NAME" ]; then
-    echo "No binary found: /app/$CHAL_NAME"
-    exit 1
-fi
-
-if [ "$CHAL_NAME" != "$DEFAULT_CHAL_NAME" ]; then
-    rm -f "/app/$DEFAULT_CHAL_NAME"
-fi
+[[ "$BASE" != "ynetd" && "$BASE" != "socat" ]] && invalid "BASE" "ynetd/socat"
+[ ! -f "/app/$CHAL_NAME" ] &&  error "No base-binary found: \e[33m/app/$CHAL_NAME\e[0m"
+[ "$CHAL_NAME" != "$DEFAULT_CHAL_NAME" ] &&  rm -f "/app/$DEFAULT_CHAL_NAME"
 
 if [ "$FLAG_FILE" != "$DEFAULT_FLAG_FILE" ]; then
     rm -f "$DEFAULT_FLAG_FILE"
     # Generate the symlink if `$FLAG_FILE_SYMLINK` is set.
-    [ ! -z "$FLAG_FILE_SYMLINK" ] && ln -s "$FLAG_FILE" "$DEFAULT_FLAG_FILE"
+    [ ! -z "$FLAG_FILE_SYMLINK" ] && (ln -s "$FLAG_FILE" "$DEFAULT_FLAG_FILE"; debug "Creating a symbolic link of $FLAG_FILE to $DEFAULT_FLAG_FILE")
 fi
 
-# self-yeet
-rm -- "$0"
+# Check if the running-container is a python container:
+if [[ "$1" == "IS_PY" ]]; then
+    debug "Checking if /app/$CHAL_NAME" contains the shebang
+    # Check if the first line of `/app/$CHAL_NAME` is not a shebang; add it:
+    FIRSTLINE=$(head -n 1 "/app/$CHAL_NAME")
+    if [ ! "${FIRSTLINE:0:3}" == '#!/' ]; then
+        (echo '#!/usr/bin/env python3' | cat - "/app/$CHAL_NAME") > tmp && mv tmp "/app/$CHAL_NAME"
+    fi
+    INVOKE=python3
+fi
 
 # Setting the permissions 550 on the /app/CHAL_NAME and 440 on flag
 # Since we're not sure the flag is in / or /app (I sometimes add in / as well)
 # So, I'm going to go for a wildcard:
+chown root:$RUN_AS /app/$CHAL_NAME "$FLAG_FILE" /flag* &>/dev/null
 chmod 550 "/app/$CHAL_NAME"
 chmod 440 "$FLAG_FILE" /flag* &>/dev/null
 
@@ -73,23 +119,20 @@ fi
 chattr +i "$FLAG_FILE" "/app/$CHAL_NAME" &>/dev/null 
 
 ###### QEMU SETUP #######
-
-if [ -z "$LIBRARY_PATH" ]; then
-    echo "LIBRARY_PATH not set!"
-    exit 1
-fi
-
-if [ -z "$EMULATOR" ]; then
-    echo "EMULATOR not set!"
-    exit 1
-fi
+[ -z "$LIBRARY_PATH" ] && error "LIBRARY_PATH not set!"
+[ -z "$EMULATOR" ] && error "EMULATOR not specified!"
 
 ln -s "$LIBRARY_PATH/lib/ld-linux-aarch64.so.1" "/usr/lib/ld-linux-aarch64.so.1" &>/dev/null
-ln -s "$LIBRARY_PATH/lib/ld-linux.so.3" "/lib/ld-linux.so.3" &>/dev/null
 ln -s "$LIBRARY_PATH/lib/libc.so.6" "/lib/libc.so.6" &>/dev/null
+if [[ "$EMULATOR" == "qemu-arm" ]]; then
+    ln -s "$LIBRARY_PATH/lib/ld-linux-armhf.so.3" "/lib/ld-linux-armhf.so.3" &>/dev/null
+else
+    ln -s "$LIBRARY_PATH/lib/ld-linux.so.3" "/lib/ld-linux.so.3" &>/dev/null
+    ln -s /usr/lib/ld-linux-aarch64.so.1 "/lib/ld-linux-aarch64.so.1" &>/dev/null
+fi
 
 export LD_LIBRARY_PATH="$LIBRARY_PATH"
-echo "[QEMU] using $EMULATOR and libaries @ $LIBRARY_PATH"
+info "[\e[33mQEMU\e[0m] using \e[32m$EMULATOR\e[0m and libaries @ \e[36m$LIBRARY_PATH\e[0m"
 
 cd "$START_DIR";
 
@@ -97,17 +140,17 @@ cd "$START_DIR";
 ## NOTE: The binary will be hosted, but the stdin/out/err won't be redirected, so you'll have to communicate using docker :(
 # if someone can help me with this, that'll be helpful, ty.
 if [ ! -z "$QEMU_GDB_PORT" ]; then
-    echo "[GDB] Enabling QEMU's GDB remote debugging on $QEMU_GDB_PORT"
+    info "[\e[33mGDB\e[0m] Enabling QEMU's GDB remote debugging on \e[36m$QEMU_GDB_PORT\e[0m"
     while [[ 1 ]]; do
         LD_LIBRARY_PATH="$LD_LIBRARY_PATH" $EMULATOR -g $QEMU_GDB_PORT -L "$LIBRARY_PATH" "/app/$CHAL_NAME"
     done
     exit 0
 fi
 
-echo "Running $CHAL_NAME in $(pwd) as $RUN_AS using $BASE and listening locally on $PORT"
+info "Running \e[33m$CHAL_NAME\e[0m in \e[32m$(pwd)\e[0m as \e[36m$RUN_AS\e[0m using \e[35m$BASE\e[0m and listening locally on \e[34m$PORT\e[0m using \e[33m$EMULATOR\e[0m"
 if [ "$BASE" == "socat" ]; then
     rm -f /opt/ynetd
-    LD_LIBRARY_PATH="$LD_LIBRARY_PATH" $EMULATOR -L "$LIBRARY_PATH" su $RUN_AS -c "/opt/socat tcp-l:$PORT,reuseaddr,fork, EXEC:\"/app/$CHAL_NAME\",stderr | tee -a $LOG_FILE"
+    LD_LIBRARY_PATH="$LD_LIBRARY_PATH" $EMULATOR -L "$LIBRARY_PATH" /bin/su $RUN_AS -c "/opt/socat tcp-l:$PORT,reuseaddr,fork, EXEC:\"/app/$CHAL_NAME\",stderr | tee -a $LOG_FILE"
 else
     rm -f /opt/socat
     # -lt => cpu time in seconds. Keeps connection opened for max 10 seconds.
